Moved all types to dispatch_types.go && remove panic
